sinf

+++ o_f.c
@@ -16,7 +16,7 @@ int main()
 #pragma acc parallel loop vector vector_length(160) gang present(arr)
 	for (int i = 0; i < 10000000; ++i)
 	{
-		arr[i] = sin(i * step);
+		arr[i] = sinf(i * step);
 	}
 
 #pragma acc parallel loop reduction(+:sum) present(arr)
